LFG system beta v1 and filter tools
LFG system lets users find parties easier while filter enhances the toxicity detection.

+++ Events/Guild/guildDelete.js
@@ -4,11 +4,14 @@ const fs = require('fs');
 module.exports = {
     name: 'guildDelete', // executes when a server is deleted or the bot is removed from it.
     async execute(guild) {
+        if(!guild || guild.name === undefined) return;
         // remove all guild related data from the database
         const dbTablesObject = JSON.parse(fs.readFileSync('./objects/database-default-tables.json'));
         let tableNames = dbTablesObject["table_names"];
         tableNames.forEach(async (table) => {
-            await poolConnection.query(`DELETE FROM ${table} WHERE guild=$1`, [guild.id]);
+            try{
+                await poolConnection.query(`DELETE FROM ${table} WHERE guild=$1`, [guild.id]);
+            } catch(err) { console.error(`guildDelete failed removing guild from ${table} table`) };
         });
     }
 }
